bmp_logo: support CONFIG_DM_VIDEO
in case of bmp_logo, the video_bmp is used for
drawing a bmp logo. This driver supports only
"full" bmp data, so adding a logo with the bmp_logo
tool to u-boot binary does not work with it,
as bmp_tool removes header.

This patch adds now the full bmp data to the
u-boot binary, so video_bmp works again.

Signed-off-by: Heiko Schocher <hs@denx.de>