crypto, caam: disable hardware acceleration
The FSL-CAAM driver requires DMA-allocated buffers for input and output,
which nobody thought about when committing this driver.
This breaks autoboot.c as well as cmd_hash.
On the console you see:
   Error: Address arguments are not aligned
   CAAM was not setup properly or it is faulty

Disable the HW_ACCEL feature for now, u-boot uses software functions in that case.
The boot-image check always use the software functions and there for will not fail.

Signed-off-by: Heiko Schocher <hs@denx.de>