env: split env_import_redund() into 2 functions
split from env_import_redund() the part which checks
which Environment is valid into a separate function
called env_check_redund() and call it from env_import_redund().

So env_check_redund() can be used from places which also
need to do this checks.

Signed-off-by: Heiko Schocher <hs@denx.de>

Commit-notes:
The return values from env_check_redund() for the cases
only one read buffer is OK are not perfect, may we should
find better return codes?
END

Series-to: u-boot
Series-to: Simon Goldschmidt <simon.k.r.goldschmidt@gmail.com>

Series-changes: 2
- patch "env: split env_import_redund into 2 functions"
  is new in version 2. Idea is to not duplicate too much
  code as Simon suggested.

Cover-letter:
env: Access Environment in SPI flashes before relocation
Add the possibility to access Environment stored in SPI NOR
Flash before relocation. This is used in DM/DTS rework for
the aristainetos boards. There mutliple DTS are packed into
a FIT image and through a variable in the Environment
the correct DTS will be selected. For this we need very
early access to the Environment. Also may this is good
for selecting the console baudrate again through Environment
which I think currently does not work for any board based on
DM/DTS and with Environment in SPI NOR flash.

Add this to aristainetos board, as aristainetos DM board
updates are now in mainline.

travis build:
END