test_sleep.py: make sleep time and margin configurable
make the sleep time and the margin configurable.

Signed-off-by: Heiko Schocher <hs@denx.de>
Series-to: u-boot
Patch-cc: Stephen Warren <swarren@nvidia.com>
Patch-cc: Simon Glass <sjg@chromium.org>

Commit-notes:
This is needed as I use tests now within tbot. On some
configurations U-Boot compiled on a build machine for example
in munich, while the board under test is in my lab in hungary.
So the 0.25 seconds default margin is often to low because
of latencies on the net.

END