[RFC] CI: add U-Boot testing on real HW with tbot
add U-Boot tests on real hardware with tbot, which is
started from gitlab CI (no idea where it runs).

Currently one board is implemented (imx8qxp based cxg3
board).

This is a good example, as we need to get a working binary,
some binary blobs for the build. This blobs are stored in
the tftp directory on the lab host, to which the hardware
is connected to. Tbot copies them from the lab host to
the CI host. Also we have some pending patches to get
ethernet and USB SDP bootmode up and working, which can
be easy loaded.

After build has finished, tbot copies the resulting
binaries to the lab host where with, the help of USB SDP
bootmode and uuu tool, the new flash.bin image is booted,
and with USB fastboot mode flashed into the emmc. After
that worked, bootmode is switched to emmc and the new
flash.bin gets bootet (and tbot checks the bootmode output,
so we can be sure, that really the new image boots from
the correct bootdevice). A simple tbot testcase for a ping
over ethernet is done and ut command is called.

After that the test/py framework is started (from the CI
host), no external scripts are necessary for this, all
"scripts" needed for this framework are created from tbot
at runtime, and tbot functions as a multiplexer for this
tests.

Signed-off-by: Heiko Schocher <hs@nabladev.com>
Series-to: u-boot
Series-version: 1
Patch-cc: Tom

Cover-letter:
This series introduces U-Boot tests with tbot on real hardware.

Currently only one board in my lab@hungary is implemented, if series
gets accepted more will follow. May others are interested too,
and prepare also such setups ... or simply send me hardware, I can
add to my setup.

find a README here (contained in this patch):
https://source.denx.de/u-boot/custodians/u-boot-i2c/-/blob/tbottesting/tbottesting/README.md?ref_type=heads

Current board is the imx8qxp based cxg3 board. To build and get a
working image you need some binary blobs, which are copied from the
lab host to the "CI host". After successfull build the resulting
images are copied to the lab host, and new flash.bin is booted
with uuu tool in USB SDP bootmode, and than installed with fastboot
to the emmc. Than power is switched off, bootmode set to emmc
and it gets checked, that U-Boot boots from emmc now.

Than test/py gets configured and started on the CI host which
runs the tests on the board in tha lab.

Hopefully the approach is generic enough to get other hardware
easy integrated, but if there are problems, just contact me,
if stuff is missing.

You can find a running test example @gitlab in my i2c custodian branch
here:

https://source.denx.de/u-boot/custodians/u-boot-i2c/-/jobs/1081203
(If you have no access rights, look below for a test started @github)

I dropped from above repo/branch the following commits in this series:

- tbottest: add github example
  as it is not interesting for U-Boot mainline (or?)

  Just for interest, with that the same test is started from repo @github, example:
  https://github.com/hsdenx/u-boot-test/actions/runs/14165226208/job/39677157779

- .gitlab-ci.yml: only start tbot tests
- remove .azure-pipelines.yml
  as this changes are only for speeding up my local testing.

The patches in this series:

- test/py/requirements.txt: update coverage
- tools/buildman/requirements.txt: update coverage
- test/py/requirements.txt: increase filelock version
- test/py/requirements.txt: remove unittest2 and testtools
- test: increase timeouts

are fixing small common issues with test/py setup ...
which could be go seperate into mainline, but in the first
shot, I want to have all patches in one series.

- CI: add U-Boot testing on real HW with tbot

  adds all the code needed for testing the imx8qxp based cxg3
  board with tbot in tbottest subdirectory. We should discuss where
  to put this code!

  May into u-boot:/test/tbottest ?

  This setup not only runs on gitlab, it also runs from github,
  jenkins ... or from a local shell without any changes in code. So you
  can use it for your normal development process (what is the goal from
  tbot) as well without any change.

- .gitlab-ci.yml: add tbot-lab

  finally integrates the tbot test setup into .gitlab-ci.yml

  This works currently for my I2C custodian repo, with some CI secrets
  stored... we have to add this secrets to u-boot repo to get it up
  and running.

  If this is not possible may I should setup a gitlab runner?
  (Which I would like to prevent... but possible of course, and this
   would prevent the timing issues I fixed when tbot is started somewhere
   in the world)

Bonus patch with an idea for custodians I want to implement, if code
is accepted:

- scripts: add helper for getting infos from your patchwork ToDo list

  is a preparation for activating my 8 year old setup, which
  downloaded patches from my Patchwork ToDo list, checked them
  with checkpatch.pl script  and applied them with git am, fully
  automated from within the CI !

  With that script it is very easy to reactivate this now again,
  if this is interesting for others. Currently I have not a clean
  patchwork ToDo list, so I add this in a second step.

  My plan is also to reactivate an automated git bisect session,
  in case there are problems with that patches, so test shows the
  problematic patch... which I think would be a great feature and
  benefit to have this mainline.

  For reference an old video on youtube how this worked 8 years ago:

  https://www.youtube.com/watch?v=PhaYfqOrQOg

  It would be nice, we can make such a step configurable through
  a switch we can enable/disable before starting a gitlab CI run.

May time is now ready for it, that people have interests in such
a setup?

END